% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAll.R
\docType{data}
\name{cols}
\alias{cols}
\title{hsapiens column data we are targetting}
\format{
An object of class \code{DFrame} with 572 rows and 75 columns.
}
\usage{
cols
}
\description{
A dataframe containing metadata for the dataset we are processing; notably the accessions (which tell us the specific chunks of data to download from dee2.io) are in cols$SRR_accession
}
\examples{
# We can use this for looking up metadata based on other metadata
# For instance, to get all the accessions for liver cells in the dataset we are analysing, run
liver_cell_accessions <- homosapienDEE2CellScore::cols$SRR_accession[homosapienDEE2CellScore::cols$cell_type == "liver"]
}
\keyword{datasets}
