% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAll.R
\name{downloadAllTheData}
\alias{downloadAllTheData}
\title{downloadAllTheData in SummarizedExperiment format}
\usage{
downloadAllTheData()
}
\description{
This is a helper function to download all of the processed data from figshare and unpack it into a tagged list of SummarizedExperiment objects.
}
\examples{
# To download all of the preprocessed data from figshare via ExperimentHub, run:

#the_data <- homosapienDEE2CellScore::downloadAllTheData()
}
