% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAll.R
\name{readInSEFolder}
\alias{readInSEFolder}
\title{readInSEFolder read a SummarizedExperiment in from a folder}
\usage{
readInSEFolder(folder_name = "SE_out/")
}
\arguments{
\item{folder_name}{The path to a folder containing a SummarizedExperiment}
}
\value{
A SummarizedExperiment object.
}
\description{
This function reads in a SummaraizedExperiment from a Zip file generated by writeOutSEZip.
It is designed for getting an intact SummarizedExperiment out of ExperimentHub for a data package,
so it does not clean up after itself and leaves stray csv files in the data package directory.
}
\examples{
# We can read in a small SummarizedExperiment containing a
# subset of the built data stored directly in the package like so
small_data <- readInSEFolder(
  folder_name=system.file("ExampleSummarisedExperimentFolder", package="homosapienDEE2CellScore"))
}
